feat: db state sync by merk chunking (#292)
* wip

    finalize multi chunk with limit

    implement chunk op

    test chunk op encoding

    convert chunk op chunk id to string

    implement traversal instruction to string

    add chunking error + devoid multi subtree chunk from encoding work

    multi-subtree-chunk skeleton + return encoding length in multichunk

    make chunks fixed size height 2

    test height proof

    implement height proof verifier

    update documentation

    verify height proof generation

    add documentation

    test no of chunk under chunk id fn

    implement number of chunks under chunk id function

    extract chunk layer function from chunk height

    seperate number_of_chunk into height and layer_height functions

    return multi chunk result

    enforce limit without storage overhead

    add test for encoding length check

    implement iterator for chunk producer

    remove cost from chunks

    fix the error type

    implement random chunk access

    fixes

    implement chunk height function

    add traverse then build chunk function to ref walker

    update comment

    implement chunk producer length

    init chunk producer struct

    implement merk tree height function

    update traversal generation instruction

    add instruction traversal test

    fix documentation

    implement binary range function

    clean up number of chunks function

    given a subtree of a given height return the exit node count

    documentation fixes

    implement chunk_height_per_layer

    verify that chunks produce expected root hash

    implement and test variable depth chunk creation

    restart chunking v2

* Squashed commit of the following:

    remove bad test

    rename files

    update documentation

    wip

    wip

    implement merk verifier + state building

    implement replication from multichunk

    fix chunk verification

    fixed implementation of chunkid from traversal instructions

    fix some tests

    make chunk_id from traversal instruction test resistant to changes in underlying chunking scheme

    add restoration logic test function

    returning the next chunk id when you call chunk

    use strings as communication interface between producer and restorer

    implement chunk id from traversal instruction

    add traversal instruction generation to direct string

    chunk producer returns next index as string for multi chunk

    clean up rewrite parent links

    restoration done successfully

    rough implementation of rewrite parent

    implement function to extract sum from node type

    wip

    chunk write logic + restorer finalization + parent key tracking

    new visit ref function that keeps track of traversal path

    implement instruction string to traversal instruction

    test child to link functionality for basic and sum merks

    implement node to link include sum

    wip

    implement and test chunk verification

    Fix layer iter function

    Previous implementation made a key assumption that nodes are unique
    including hash nodes, this made the layer iteration functionality
    depend on the contents of the tree, which shouldn't be the case.

    This adds a simpler implementation of the layer iter logic using breadth
    first search.

    add test to ensure chunks only contain hash and kvfeaturetype

    test for avl tree during proof op execution

    remove chunk_height_per_layer_lin_comb every chunk now has fixed height of 2

* wip

* wip

* wip

* rename job

* clippy fixes

* feat: base state sync

* dynamic chunk id calculation

* more work

* more work

* more work

* fix: help with lifetimes

* more work

* final work

* cargo fmt

* more fmt

* clippy fixes

* more fmt

* fix for verify feature

* more fmt

* test fixes

* more fmt

* refactor

* refactor

* refactor

* more refactoring

* more refactoring

* fmt

* suggestions

---------

+++ grovedb/src/tests/query_tests.rs
@@ -46,7 +46,7 @@ use crate::{
 fn populate_tree_for_non_unique_range_subquery(db: &TempGroveDb) {
     // Insert a couple of subtrees first
     for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
+        let i_vec = i.to_be_bytes().to_vec();
         db.insert(
             [TEST_LEAF].as_ref(),
             &i_vec,
@@ -70,7 +70,7 @@ fn populate_tree_for_non_unique_range_subquery(db: &TempGroveDb) {
 
         for j in 100u32..150 {
             let mut j_vec = i_vec.clone();
-            j_vec.append(&mut (j as u32).to_be_bytes().to_vec());
+            j_vec.append(&mut j.to_be_bytes().to_vec());
             db.insert(
                 [TEST_LEAF, i_vec.as_slice(), b"\0"].as_ref(),
                 &j_vec.clone(),
@@ -87,7 +87,7 @@ fn populate_tree_for_non_unique_range_subquery(db: &TempGroveDb) {
 fn populate_tree_for_non_unique_double_range_subquery(db: &TempGroveDb) {
     // Insert a couple of subtrees first
     for i in 0u32..10 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
+        let i_vec = i.to_be_bytes().to_vec();
         db.insert(
             [TEST_LEAF].as_ref(),
             &i_vec,
@@ -110,7 +110,7 @@ fn populate_tree_for_non_unique_double_range_subquery(db: &TempGroveDb) {
         .expect("successful subtree insert");
 
         for j in 25u32..50 {
-            let j_vec = (j as u32).to_be_bytes().to_vec();
+            let j_vec = j.to_be_bytes().to_vec();
             db.insert(
                 [TEST_LEAF, i_vec.as_slice(), b"a"].as_ref(),
                 &j_vec,
@@ -134,7 +134,7 @@ fn populate_tree_for_non_unique_double_range_subquery(db: &TempGroveDb) {
             .expect("successful subtree insert");
 
             for k in 100u32..110 {
-                let k_vec = (k as u32).to_be_bytes().to_vec();
+                let k_vec = k.to_be_bytes().to_vec();
                 db.insert(
                     [TEST_LEAF, i_vec.as_slice(), b"a", &j_vec, b"\0"].as_ref(),
                     &k_vec.clone(),
@@ -173,7 +173,7 @@ fn populate_tree_by_reference_for_non_unique_range_subquery(db: &TempGroveDb) {
     .expect("successful subtree insert");
     // Insert a couple of subtrees first
     for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
+        let i_vec = i.to_be_bytes().to_vec();
         db.insert(
             [TEST_LEAF, b"1"].as_ref(),
             &i_vec,
@@ -198,7 +198,7 @@ fn populate_tree_by_reference_for_non_unique_range_subquery(db: &TempGroveDb) {
         for j in 100u32..150 {
             let random_key = rand::thread_rng().gen::<[u8; 32]>();
             let mut j_vec = i_vec.clone();
-            j_vec.append(&mut (j as u32).to_be_bytes().to_vec());
+            j_vec.append(&mut j.to_be_bytes().to_vec());
 
             // We should insert every item to the tree holding items
             db.insert(
@@ -231,7 +231,7 @@ fn populate_tree_by_reference_for_non_unique_range_subquery(db: &TempGroveDb) {
 fn populate_tree_for_unique_range_subquery(db: &TempGroveDb) {
     // Insert a couple of subtrees first
     for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
+        let i_vec = i.to_be_bytes().to_vec();
         db.insert(
             [TEST_LEAF].as_ref(),
             &i_vec,
@@ -278,7 +278,7 @@ fn populate_tree_by_reference_for_unique_range_subquery(db: &TempGroveDb) {
     .expect("successful subtree insert");
 
     for i in 1985u32..2000 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
+        let i_vec = i.to_be_bytes().to_vec();
         db.insert(
             [TEST_LEAF, b"1"].as_ref(),
             &i_vec,
@@ -333,7 +333,7 @@ fn populate_tree_for_unique_range_subquery_with_non_unique_null_values(db: &mut 
     .expect("successful subtree insert");
     // Insert a couple of subtrees first
     for i in 100u32..200 {
-        let i_vec = (i as u32).to_be_bytes().to_vec();
+        let i_vec = i.to_be_bytes().to_vec();
         db.insert(
             [TEST_LEAF, &[], b"\0"].as_ref(),
             &i_vec,
